yt_video_transcript_api/requirements.txt
fastapi[all]
uvicorn
youtube_transcript_api
requests
yt_dlp
langchain_community
deepgram-sdk